[py3] Add loading support for py2 and py3 interop
The payloads that come in should normally be from the same runtime as
the source, but in the transition phase that is not true.  This adds an
exception handling section that will use a fallback loader (using a
bytes encoder) and then a normalization function to convert any
bytestrings into native strings.  Again, this should only be a path
reached during the initial transition phase and then the except clause
should go dark...